moved updload to scan save
