+++ backend/app/core/scanner_bl.py
@@ -1,5 +1,9 @@
 from util.serializer import deserialize, serialize
 from core.scanner_db import save_scan_db
+from fastapi import APIRouter, File, UploadFile
+import datetime, os, uuid
+
+IMAGEDIR = "images/"
 
 cv_rs_dummy_json = '{"candidate":{"number":"CHSG-23.123","date_of_birth":"2010-01-01"},"exams":[{"year":2023,' \
                    '"subject":"ABC English","total_score":3.75,"exercises":[{"number":"1.a","score":1.75,' \
@@ -8,14 +12,14 @@ cv_rs_dummy_json = '{"candidate":{"number":"CHSG-23.123","date_of_birth":"2010-0
                    '"score":5.00,"accuracy":0.95}]}]}'
 
 
-def save_scan():
+def save_scan(file: UploadFile = File(...)):
     """
     Save a scan by storing the file into the file storage, request extraction with cv module,
     save the data into the db and return a json.
     """
 
     # save image to file system
-
+    create_upload_file(file)
     # call computer vision
     # mock-impl. to continue implementation of response handling
     # cv_rs = DigitRecognizer().recognize_digits_in_photo(photo={})
@@ -26,3 +30,22 @@ def save_scan():
 
     # respond with json
     return serialize(db_rs)
+
+
+async def create_upload_file(file: UploadFile = File(...)):
+    file.filename = f"{uuid.uuid4()}.jpg"
+    contents = await file.read()
+    #for folder get year
+    today = datetime.date.today()
+    year = today.year
+    path = f"{IMAGEDIR}{year}"
+    exists = os.path.exists(path)
+    if not exists:
+        if not os.path.exists(IMAGEDIR):
+            os.mkdir(IMAGEDIR)
+        os.mkdir(path)
+    # save the file
+    with open(f"{IMAGEDIR}/{year}/{file.filename}", "wb") as f:
+        f.write(contents)
+# TODO: bild als base64 an tobis klasse übergeben
+    return {"filename": file.filename}
